+++ packages/core-js/es/array-buffer/constructor.js
@@ -1,4 +1,5 @@
 require('../../modules/es.array-buffer.constructor');
+require('../../modules/es.array-buffer.slice');
 require('../../modules/es.object.to-string');
 var path = require('../../internals/path');