+++ packages/core-js/features/structured-clone.js
@@ -1,0 +1,3 @@
+var parent = require('../stable/structured-clone');
+
+module.exports = parent;