+++ tests/compat/tests.js
@@ -1671,6 +1671,10 @@ GLOBAL.tests = {
   'web.queue-microtask': function () {
     return Object.getOwnPropertyDescriptor(GLOBAL, 'queueMicrotask').value;
   },
+  'web.structured-clone': function () {
+    var test = structuredClone(new AggregateError([1], 'a', { cause: 3 }));
+    return test.name == 'AggregateError' && test.errors[0] == 1 && test.message == 'a' && test.cause == 3;
+  },
   'web.timers': function () {
     return !/MSIE .\./.test(USERAGENT);
   },